refactor: add animated pressable for annoncement card
